Add existing file

+++ GlowOn/GlowOn/ViewModels/Pages/ProductsCatalogueViewModel.cs
@@ -7,6 +7,9 @@ using GlowOn.Views;
 using GlowOn.Models;
 using GlowOn.ViewModels.Interfaces;
 using Xamarin.Forms;
+using System.Collections;
+using System.Collections.Generic;
+using GlowOn.Views.Products;
 
 namespace GlowOn.ViewModels.Pages
 {
@@ -17,10 +20,9 @@ namespace GlowOn.ViewModels.Pages
         private IPageService _pageService;
 
         private bool _isDataLoaded;
-
         public ObservableCollection<ProductViewModel> Products { get; private set; }
             = new ObservableCollection<ProductViewModel>();
-
+        public IList<Product> ProductsL { get { return ProductData.Products; } }
         public ProductViewModel SelectedProduct
         {
             get { return _selectedProduct; }
@@ -31,7 +33,6 @@ namespace GlowOn.ViewModels.Pages
         public ICommand AddProductCommand { get; private set; }
         public ICommand SelectProductCommand { get; private set; }
         public ICommand DeleteProductCommand { get; private set; }
-        public ICommand SearchCommand { get; private set; }
 
         public ProductsCatalogueViewModel(IProductCatalogue productCatalogue, IPageService pageService)
         {
@@ -43,19 +44,19 @@ namespace GlowOn.ViewModels.Pages
             SelectProductCommand = new Command<ProductViewModel>(async p => await SelectProduct(p));
             DeleteProductCommand = new Command<ProductViewModel>(async p => await DeleteProduct(p));
 
-            MessagingCenter.Subscribe<ProductsDetailViewModel, Product>
+            MessagingCenter.Subscribe<ProductsEditorViewModel, Product>
                 (this, Events.ProductAdded, OnProductAdded);
 
-            MessagingCenter.Subscribe<ProductsDetailViewModel, Product>
+            MessagingCenter.Subscribe<ProductsEditorViewModel, Product>
                 (this, Events.ProductUpdated, OnProductUpdated);
         }
 
-        private void OnProductAdded(ProductsDetailViewModel source, Product product)
+        private void OnProductAdded(ProductsEditorViewModel source, Product product)
         {
             Products.Add(new ProductViewModel(product));
         }
 
-        private void OnProductUpdated(ProductsDetailViewModel source, Product product)
+        private void OnProductUpdated(ProductsEditorViewModel source, Product product)
         {
             var productInList = Products.Single(p => p.id == product.ID);
 
@@ -67,7 +68,8 @@ namespace GlowOn.ViewModels.Pages
             productInList.Category = product.Category;
             productInList.Ingredients = product.Ingredients;
             productInList.AgeCategory = product.AgeCategory;
-            productInList.ExtraInfo = product.ExtraInfo;
+            productInList.SuggestedPrice = product.SuggestedPrice;
+            productInList.IsInitial = product.IsInitial;
         }
 
         private async Task LoadData()
@@ -76,6 +78,12 @@ namespace GlowOn.ViewModels.Pages
                 return;
 
             _isDataLoaded = true;
+
+            foreach (Product p in ProductsL)
+            {
+                await _productCatalogue.AddOrReplace(p);
+            }
+
             var products = await _productCatalogue.GetProductsAsync();
             foreach (var product in products)
                 Products.Add(new ProductViewModel(product));
@@ -83,7 +91,7 @@ namespace GlowOn.ViewModels.Pages
 
         private async Task AddProduct()
         {
-            await _pageService.PushAsync(new ProductsDetail(new ProductViewModel()));
+            await _pageService.PushAsync(new ProductsEditor(new ProductViewModel()));
         }
 
         private async Task SelectProduct(ProductViewModel product)
@@ -92,15 +100,22 @@ namespace GlowOn.ViewModels.Pages
                 return;
 
             SelectedProduct = null;
-            await _pageService.PushAsync(new ProductsDetail(product));
+            if (product.IsInitial == true)
+                await _pageService.PushAsync(new ProductsDetail(product));
+            else
+                await _pageService.PushAsync(new ProductsEditor(product));
         }
 
         private async Task DeleteProduct(ProductViewModel productViewModel)
         {
+            if(productViewModel.IsInitial)
+            {
+                await _pageService.DisplayAlert("Error", "Your cannot delete a default product.", "OK");
+                return;
+            }
             if (await _pageService.DisplayAlert("Warning", $"Are you sure you want to delete {productViewModel.ProductName}?", "Yes", "No"))
             {
                 Products.Remove(productViewModel);
-
                 var product = await _productCatalogue.GetProduct(productViewModel.id);
                 await _productCatalogue.DeleteProduct(product);
             }
